clean up methods, migration, and add comment for mutation
