+++ server/graphql/vvm/src/mutations/vvm_status_log/update.rs
@@ -49,6 +49,8 @@ pub fn update_vvm_status_log(
     store_id: &str,
     input: UpdateVVMStatusLogInput,
 ) -> Result<UpdateVVMStatusResponse> {
+    // TODO: Change this to proper VVM-specific permission
+    // Add "View and edit vaccine vial monitor status" user permission from OG
     let user = validate_auth(
         ctx,
         &ResourceAccessRequest {
