+++ server/service/src/vvm/vvm_status_log/mod.rs
@@ -1,3 +1,4 @@
 pub mod insert;
 pub mod query;
 pub mod update;
+pub mod validate;
