+++ server/service/src/vvm/vvm_status_log/update/validate.rs
@@ -1,20 +1,16 @@
-use repository::{
-    vvm_status::{
-        vvm_status_log_row::{VVMStatusLogRow, VVMStatusLogRowRepository},
-        vvm_status_row::VVMStatusRowRepository,
-    },
-    RepositoryError,
-};
+use repository::vvm_status::vvm_status_log_row::VVMStatusLogRow;
 
 use super::{UpdateVVMStatusLogError, UpdateVVMStatusLogInput};
-use crate::StorageConnection;
+use crate::{
+    vvm::vvm_status_log::validate::{check_vvm_status_exists, get_vvm_status_log},
+    StorageConnection,
+};
 
 pub fn validate(
     input: &UpdateVVMStatusLogInput,
     connection: &StorageConnection,
 ) -> Result<VVMStatusLogRow, UpdateVVMStatusLogError> {
-    let vvm_status_log = VVMStatusLogRowRepository::new(connection).find_one_by_id(&input.id)?;
-
+    let vvm_status_log = get_vvm_status_log(&input.id, connection)?;
     let vvm_status_log = match vvm_status_log {
         Some(vvm_status_log) => vvm_status_log,
         None => return Err(UpdateVVMStatusLogError::VVMStatusLogDoesNotExist),
@@ -27,12 +23,4 @@ pub fn validate(
     }
 
     Ok(vvm_status_log)
-}
-
-pub fn check_vvm_status_exists(
-    status_id: &str,
-    connection: &StorageConnection,
-) -> Result<bool, RepositoryError> {
-    let vvm_status = VVMStatusRowRepository::new(connection).find_one_by_id(status_id)?;
-    Ok(vvm_status.is_some())
 }
